+++ packages/datadog_flags/lib/src/flags_configuration.dart
@@ -10,15 +10,21 @@ import 'datadog_context.dart';
 class DatadogFlagsConfiguration {
   final Uri? customFlagsEndpoint;
   final Map<String, String>? customFlagsHeaders;
+  final Uri? customExposureEndpoint;
+  final bool trackExposures;
   final http.Client? httpClient;
   final DatadogFlagsContext? datadogContext;
   final bool rumIntegrationEnabled;
+  final DateTime Function() dateProvider;
 
   const DatadogFlagsConfiguration({
     this.customFlagsEndpoint,
     this.customFlagsHeaders,
+    this.customExposureEndpoint,
+    this.trackExposures = true,
     this.httpClient,
     this.datadogContext,
     this.rumIntegrationEnabled = true,
+    this.dateProvider = DateTime.now,
   });
 }